requirements.txt
# Core deep learning
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0

# Computer vision and image processing
Pillow>=8.3.0
opencv-python>=4.5.0

# NLP and evaluation metrics
nltk>=3.6.0
rouge-score>=0.1.2

# Progress bars and utilities
tqdm>=4.62.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Optional: For video/GIF generation
imageio>=2.9.0
imageio-ffmpeg>=0.4.0

# Optional: For CIDEr metric (requires Java)
# git+https://github.com/tylin/coco-caption.git

# Data handling
pandas>=1.3.0
scikit-learn>=0.24.0

# Configuration
pyyaml>=5.4.0

# Jupyter notebook (for experiments)
jupyter>=1.0.0
ipykernel>=6.0.0
